Changed data type of totalPaid from Float to BigDecimal

+++ auction-platform/src/main/java/com/eecs4413/auction_platform/dto/ReceiptResponseDTO.java
@@ -19,7 +19,7 @@ public class ReceiptResponseDTO {
      private String city;
      private String country;
      private String postalCode;
-     private Float totalPaid;
+     private BigDecimal totalPaid;
      private Long ItemID;
      private OffsetDateTime shippingDate;
      private String message;
